SUMO-103799: refactor RealisticSearchClient
