+++ elasticsearch-core/src/main/scala/com/sumologic/elasticsearch/restlastic/RestlasticSearchClient.scala
@@ -19,15 +19,28 @@
 package com.sumologic.elasticsearch.restlastic
 
 import com.sumologic.elasticsearch.restlastic.RestlasticSearchClient.ReturnTypes.{ScrollId, SearchResponse}
-import com.sumologic.elasticsearch.restlastic.dsl.{Dsl, EsVersion}
+import com.sumologic.elasticsearch.restlastic.dsl.EsVersion
 import org.json4s.FieldSerializer.{renameFrom, renameTo}
-import org.json4s._
 import org.json4s.jackson.JsonMethods._
+import scala.concurrent.Await
+
+import akka.actor.ActorSystem
+import akka.io.IO
+import akka.pattern.ask
+import akka.util.Timeout
+import com.sumologic.elasticsearch.restlastic.dsl.Dsl
+import org.json4s._
+import org.slf4j.LoggerFactory
+import spray.can.Http
 import spray.http.HttpMethods._
 import spray.http.Uri.{Query => UriQuery}
+import spray.http.HttpResponse
 import spray.http._
 
+import scala.concurrent.duration._
 import scala.concurrent.{ExecutionContext, Future}
+
+
 
 trait ScrollClient {
 
@@ -68,12 +81,18 @@ trait RequestSigner {
   def withAuthHeader(httpRequest: HttpRequest): HttpRequest
 }
 
-abstract class RestlasticSearchClient(searchExecutionCtx: ExecutionContext) extends ScrollClient {
+abstract class RestlasticSearchClient(endpointProvider: EndpointProvider, signer: Option[RequestSigner],
+                                      override val indexExecutionCtx: ExecutionContext,
+                                      searchExecutionCtx: ExecutionContext)
+                                     (implicit val system: ActorSystem = ActorSystem(),
+                                      val timeout: Timeout = Timeout(30 seconds)) extends ScrollClient {
+
+  protected val logger = LoggerFactory.getLogger(RestlasticSearchClient.getClass)
 
   import Dsl._
   import RestlasticSearchClient.ReturnTypes._
 
-  def ready: Boolean
+  def ready: Boolean = endpointProvider.ready
 
   def query(index: Index,
             tpe: Type,
@@ -89,56 +108,164 @@ abstract class RestlasticSearchClient(searchExecutionCtx: ExecutionContext) exte
     }
   }
 
-  def bucketNestedAggregation(index: Index, tpe: Type, query: AggregationQuery): Future[BucketNested]
+  def bucketNestedAggregation(index: Index, tpe: Type, query: AggregationQuery): Future[BucketNested] = {
+    implicit val ec = searchExecutionCtx
+    runEsCommand(query, s"/${index.name}/${tpe.name}/_search").map { rawJson =>
+      BucketNested(rawJson.mappedTo[BucketNestedAggregationResponse].aggregations._2)
+    }
+  }
 
-  def bucketAggregation(index: Index, tpe: Type, query: AggregationQuery): Future[BucketAggregationResultBody]
+  def bucketAggregation(index: Index, tpe: Type, query: AggregationQuery): Future[BucketAggregationResultBody] = {
+    implicit val ec = searchExecutionCtx
+    runEsCommand(query, s"/${index.name}/${tpe.name}/_search").map { rawJson =>
+      rawJson.mappedTo[BucketAggregationResponse].aggregations.aggs_name
+    }
+  }
 
-  def suggest(index: Index, tpe: Type, query: SuggestRoot): Future[Map[String,List[String]]]
+  def suggest(index: Index, tpe: Type, query: SuggestRoot): Future[Map[String,List[String]]] = {
+    // I'm not totally sure why, but you don't specify the type for _suggest queries
+    implicit val ec = searchExecutionCtx
+    val fut = runEsCommand(query, s"/${index.name}/_search")
+    fut.map { resp =>
+      val extracted = resp.mappedTo[SuggestResult]
+      extracted.suggestions
+    }
+  }
 
-  def count(index: Index, tpe: Type, query: QueryRoot): Future[Int]
+  def count(index: Index, tpe: Type, query: QueryRoot): Future[Int] = {
+    implicit val ec = searchExecutionCtx
+    val fut = runEsCommand(query, s"/${index.name}/${tpe.name}/_count")
+    fut.map(_.mappedTo[CountResponse].count)
+  }
 
-  def index(index: Index, tpe: Type, doc: Document): Future[IndexResponse]
+  def index(index: Index, tpe: Type, doc: Document): Future[IndexResponse] = {
+    implicit val ec = indexExecutionCtx
+    runEsCommand(doc, s"/${index.name}/${tpe.name}/${doc.id}").map(_.mappedTo[IndexResponse])
+  }
 
-  def deleteById(index: Index, tpe: Type, id: String): Future[DeleteResponse]
+  def deleteById(index: Index, tpe: Type, id: String): Future[DeleteResponse] = {
+    implicit val ec = indexExecutionCtx
+    runEsCommand(NoOp, s"/${index.name}/${tpe.name}/$id", DELETE).map(_.mappedTo[DeleteResponse])
+  }
 
   def deleteByQuery(index: Index, tpe: Type, query: QueryRoot, waitForCompletion: Boolean): Future[RawJsonResponse]
 
-  def documentExistsById(index: Index, tpe: Type, id: String): Future[Boolean]
+  def documentExistsById(index: Index, tpe: Type, id: String): Future[Boolean] = {
+    implicit val ec = indexExecutionCtx
+    runEsCommand(NoOp, s"/${index.name}/${tpe.name}/$id", HEAD).map(_ => true).recover {
+      case ex: ElasticErrorResponse if ex.status == 404 =>
+        false
+    }
+  }
 
-  def bulkIndex(bulk: Bulk): Future[Seq[BulkItem]]
+  def bulkIndex(bulk: Bulk): Future[Seq[BulkItem]] = {
+    implicit val ec = indexExecutionCtx
+    runEsCommand(bulk, s"/_bulk").map { resp =>
+      val bulkResp = resp.mappedTo[BulkIndexResponse]
+      bulkResp.items.map(_.values.head)
+    }
+  }
 
-  def bulkIndex(index: Index, tpe: Type, documents: Seq[Document]): Future[Seq[BulkItem]]
+  def bulkIndex(index: Index, tpe: Type, documents: Seq[Document]): Future[Seq[BulkItem]] = {
+    val bulkOperation = Bulk(documents.map(BulkOperation(create, Some(index -> tpe), _)))
+    bulkIndex(bulkOperation)
+  }
 
   // retryOnConflictOpt specifies how many times to retry before throwing version conflict exception.
   // https://www.elastic.co/guide/en/elasticsearch/reference/2.3/docs-update.html#_parameters_2
-  def bulkUpdate(index: Index, tpe: Type, documents: Seq[Document], retryOnConflictOpt: Option[Int] = None): Future[Seq[BulkItem]]
+  def bulkUpdate(index: Index, tpe: Type, documents: Seq[Document], retryOnConflictOpt: Option[Int] = None): Future[Seq[BulkItem]] = {
+    val bulkOperation = Bulk(documents.map(BulkOperation(update, Some(index -> tpe), _, retryOnConflictOpt)))
+    bulkIndex(bulkOperation)
+  }
 
-  def bulkDelete(index: Index, tpe: Type, documents: Seq[Document]): Future[Seq[BulkItem]]
+  def bulkDelete(index: Index, tpe: Type, documents: Seq[Document]): Future[Seq[BulkItem]] = {
+    val bulkOperation = Bulk(documents.map(BulkOperation(delete, Some(index -> tpe), _)))
+    bulkIndex(bulkOperation)
+  }
 
-  def putMapping(index: Index, tpe: Type, mapping: Mapping): Future[RawJsonResponse]
+  def putMapping(index: Index, tpe: Type, mapping: Mapping): Future[RawJsonResponse] = {
+    implicit val ec = indexExecutionCtx
+    runEsCommand(mapping, s"/${index.name}/_mapping/${tpe.name}")
+  }
 
-  def getMapping(index: Index, tpe: Type): Future[RawJsonResponse]
+  def getMapping(index: Index, tpe: Type): Future[RawJsonResponse] = {
+    implicit val ec = searchExecutionCtx
+    runEsCommand(EmptyObject, s"/${index.name}/_mapping/${tpe.name}", GET)
+  }
 
   def createIndex(index: Index, settings: Option[IndexSetting] = None): Future[RawJsonResponse]
 
-  def deleteIndex(index: Index): Future[RawJsonResponse]
+  def deleteIndex(index: Index): Future[RawJsonResponse] = {
+    implicit val ec = indexExecutionCtx
+    runEsCommand(EmptyObject, s"/${index.name}", DELETE)
+  }
 
   @deprecated("When plugin is not enabled this function doesn't handle pagination, so it deletes only first page of query results. Replaced by deleteDocuments.")
-  def deleteDocument(index: Index, tpe: Type, deleteQuery: QueryRoot, pluginEnabled: Boolean = false): Future[RawJsonResponse]
+  def deleteDocument(index: Index, tpe: Type, deleteQuery: QueryRoot, pluginEnabled: Boolean = false): Future[RawJsonResponse] = {
+    implicit val ec = indexExecutionCtx
+    if (pluginEnabled) {
+      runEsCommand(deleteQuery, s"/${index.name}/${tpe.name}/_query", DELETE)
+    } else {
+      val response = Await.result(query(index, tpe, deleteQuery, rawJsonStr = false), 10.seconds).rawSearchResponse
+      val totalHits = response.hits.total
+      val documents = response.hits.hits.map(_._id)
+      if (totalHits > documents.length) {
+        logger.warn(s"deleting only first ${documents.length}/$totalHits matches. " +
+          "Use deleteDocuments, if you want to delete more at once.")
+      }
 
-  def deleteDocuments(index: Index, tpe: Type, deleteQuery: QueryRoot, pluginEnabled: Boolean = false): Future[Map[Index, DeleteResponse]]
+      bulkDelete(index, tpe, documents.map(Document(_, Map()))).map(res => RawJsonResponse(res.toString))
+    }
+  }
 
-  def scroll(scrollId: ScrollId, resultWindowOpt: Option[String] = None): Future[(ScrollId, SearchResponse)]
+  def deleteDocuments(index: Index, tpe: Type, deleteQuery: QueryRoot, pluginEnabled: Boolean = false): Future[Map[Index, DeleteResponse]] = {
+    def firstScroll(scId: ScrollId) = startScrollRequest(index, tpe, deleteQuery)
 
-  def flush(index: Index): Future[RawJsonResponse]
+    scrollDelete(index, tpe, ScrollId(""), Map.empty[Index, DeleteResponse], firstScroll)
+  }
 
-  def refresh(index: Index): Future[RawJsonResponse]
+  protected def scrollDelete(index: Index,
+                             tpe: Type,
+                             scrollId: ScrollId,
+                             acc: Map[Index, DeleteResponse],
+                             scrollingFn: (ScrollId) => Future[(ScrollId, SearchResponse)]): Future[Map[Index, DeleteResponse]]
 
-  def runRawEsRequest(op: String,
-                      endpoint: String,
-                      method: HttpMethod = POST,
-                      query: UriQuery = UriQuery.Empty)
-                     (implicit ec: ExecutionContext = ExecutionContext.Implicits.global): Future[RawJsonResponse]
+  def scroll(scrollId: ScrollId, resultWindowOpt: Option[String] = None): Future[(ScrollId, SearchResponse)] = {
+    implicit val ec = searchExecutionCtx
+    val uriQuery = UriQuery("scroll_id" -> scrollId.id, "scroll" -> resultWindowOpt.getOrElse(defaultResultWindow))
+    runEsCommand(NoOp, s"/_search/scroll", query = uriQuery).map { resp =>
+      val sr = resp.mappedTo[SearchResponseWithScrollId]
+      (ScrollId(sr._scroll_id), SearchResponse(RawSearchResponse(sr.hits), resp.jsonStr))
+    }
+  }
+
+  // Scroll requests have optimizations that make them faster when the sort order is _doc.
+  // Put sort by _doc in query as described in the the following document
+  // https://www.elastic.co/guide/en/elasticsearch/reference/current/search-request-scroll.html
+  protected def startScrollRequest(index: Index,
+                                   tpe: Type,
+                                   query: QueryRoot,
+                                   resultWindowOpt: Option[String],
+                                   fromOpt: Option[Int],
+                                   sizeOpt: Option[Int],
+                                   preference: Option[String],
+                                   params: Map[String, String]): Future[(ScrollId, SearchResponse)] = {
+    implicit val ec = searchExecutionCtx
+    runEsCommand(query, s"/${index.name}/${tpe.name}/_search", query = UriQuery(params)).map { resp =>
+      val sr = resp.mappedTo[SearchResponseWithScrollId]
+      (ScrollId(sr._scroll_id), SearchResponse(RawSearchResponse(sr.hits), resp.jsonStr))
+    }
+  }
+
+  def flush(index: Index): Future[RawJsonResponse] = {
+    implicit val ec = indexExecutionCtx
+    runEsCommand(EmptyObject, s"/${index.name}/_flush")
+  }
+
+  def refresh(index: Index): Future[RawJsonResponse] = {
+    implicit val ec = indexExecutionCtx
+    runEsCommand(EmptyObject, s"/${index.name}/_refresh")
+  }
 
   def version: EsVersion
 
@@ -154,6 +281,43 @@ abstract class RestlasticSearchClient(searchExecutionCtx: ExecutionContext) exte
       op.toJsonStr(version)
     }
     runRawEsRequest(jsonStr, endpoint, method, query)
+  }
+
+  def runRawEsRequest(op: String,
+                      endpoint: String,
+                      method: HttpMethod = POST,
+                      query: UriQuery = UriQuery.Empty)
+                     (implicit ec: ExecutionContext = ExecutionContext.Implicits.global): Future[RawJsonResponse]
+
+  protected def runRawEsRequest(op: String,
+                                endpoint: String,
+                                method: HttpMethod,
+                                query: UriQuery,
+                                request: HttpRequest)
+                               (implicit ec: ExecutionContext): Future[RawJsonResponse] = {
+    logger.debug(f"Got Rs request: $request (op was $op)")
+
+
+    val responseFuture: Future[HttpResponse] = (IO(Http) ? request)(timeout).mapTo[HttpResponse]
+
+    responseFuture.map { response =>
+      logger.debug(f"Got Es response: $response")
+      if (response.status.isFailure) {
+        logger.warn(s"Failure response: ${response.entity.asString.take(500)}")
+        logger.warn(s"Failing request: ${op.take(5000)}")
+        throw ElasticErrorResponse(JString(response.entity.asString), response.status.intValue)
+      }
+      RawJsonResponse(response.entity.asString)
+    }
+  }
+
+  protected def buildUri(path: String, query: UriQuery = UriQuery.Empty): Uri = {
+    val ep = endpointProvider.endpoint
+    val scheme = ep.port match {
+      case 443 => "https"
+      case _ => "http"
+    }
+    Uri.from(scheme = scheme, host = ep.host, port = ep.port, path = path, query = query)
   }
 }
